requirements.txt
fastapi
uvicorn
playwright
pillow>=9.0.0
runpod
setuptools
envkey
ffmpeg-python
numpy>=1.20.0